[Refactoring ⚙️] 쿼리키 전역으로 뺌
