+++ src/constants/queryKey.ts
@@ -3,4 +3,12 @@ export const QUERYKEY = {
   TECH_STACKS: "techStacks",
   ALL_PROJECTS: "allProjects",
   BANNER_PROJECTS: "bannerProjects",
+  PROJECT_DETAIL: "projectDetail",
+  DELETE_PROJECT: "deleteProject",
+
+  POST_LIKE: "postLike",
+  DELETE_LIKE: "deleteLike",
+  POST_COMMENT: "postComment",
+  EDIT_COMMENT: "editComment",
+  DELETE_COMMENT: "deleteComment",
 }
