+++ src/pages/ProjectDetailPage/hooks/mutations/useDeleteCommentMutation.ts
@@ -13,20 +13,18 @@ import {
 } from "@pages/ProjectDetailPage/constants/toastMessage"
 import { toastOptions } from "@pages/SignUpPage/constants/toastOptions"
 
-import { QUERY_KEY_GET_PROJECT_DETAIL } from "../queries/useProjectDetailQuery"
-
-const QUERY_KEY_DELETE_COMMENT = "DELETE_COMMENT_234893204832"
+import { QUERYKEY } from "@constants/queryKey"
 
 export const useDeleteCommentMutation = () => {
   const queryClient = useQueryClient()
   const toast = useToast(toastOptions)
 
   const { mutate: deleteCommentMutation, error } = useMutation({
-    mutationKey: [QUERY_KEY_DELETE_COMMENT],
+    mutationKey: [QUERYKEY.DELETE_COMMENT],
     mutationFn: (commentId: number) => deleteComment({ commentId }),
     onSuccess: () => {
       queryClient.invalidateQueries({
-        queryKey: [QUERY_KEY_GET_PROJECT_DETAIL],
+        queryKey: [QUERYKEY.PROJECT_DETAIL],
       })
     },
   })
